+++ FE/src/components/common/QuestionSelectionBox/QuestionTabPanelItem.tsx
@@ -1,18 +1,15 @@
 import { questionSetting } from '@atoms/interviewSetting';
 import { theme } from '@styles/theme';
 import { css } from '@emotion/react';
-import { useRef, useState } from 'react';
+import { useState } from 'react';
 import { useRecoilValue } from 'recoil';
 import useQuestionWorkbookQuery from '@hooks/apis/queries/useQuestionWorkbookQuery';
-import { Typography, Toggle, Tabs, CheckBox } from '@foundation/index';
+import { Typography, Toggle, Tabs } from '@foundation/index';
 import { WorkbookTitleListResDto } from '@/types/workbook';
 import { ExcludeArray } from '@/types/utils';
-import QuestionSelectionBoxAccordion from './QuestionSelectionBoxAccordion';
 import QuestionAddForm from '@common/QuestionSelectionBox/QuestionAddForm';
 import QuestionTabPanelHeader from '@common/QuestionSelectionBox/QuestionTabPanelHeader';
-import WorkbookEditModeDialog from '@common/QuestionSelectionBox/WorkbookEditModeDialog';
-import useOutsideClick from '@hooks/useOutsideClick';
-import useWorkbookQuestionDelete from '@hooks/useWorkbookQuestionDelete';
+import QuestionAccordionList from '@common/QuestionSelectionBox/QuestionAccordionList';
 
 type TabPanelItemProps = {
   selectedTabIndex: string;
@@ -35,12 +32,6 @@ const TabPanelItem: React.FC<TabPanelItemProps> = ({
   const [onlySelectedOption, setOnlySelectedOption] = useState(false);
   const [isEditMode, setIsEditMode] = useState(false);
 
-  const tabContentRef = useRef<HTMLDivElement>(null);
-
-  useOutsideClick(tabContentRef, () => {
-    handleCancelEditMode();
-  });
-
   const toggleShowSelectionOption = () => {
     setOnlySelectedOption((prev) => !prev);
   };
@@ -49,27 +40,6 @@ const TabPanelItem: React.FC<TabPanelItemProps> = ({
     workbookId: workbook.workbookId,
     enabled: selectedTabIndex === tabIndex,
   });
-  const {
-    addCheckedQuestion,
-    resetCheckedQuestion,
-    deleteCheckedQuestion,
-    isCheckedQuestion,
-    checkQuestionCount,
-  } = useWorkbookQuestionDelete(workbook.workbookId);
-
-  const handleQuestionChecked = (questionId: number) => {
-    isEditMode && addCheckedQuestion(questionId);
-  };
-
-  const handleCancelEditMode = () => {
-    resetCheckedQuestion();
-    setIsEditMode(false);
-  };
-
-  const handleDeleteQuestion = async () => {
-    await deleteCheckedQuestion();
-    handleCancelEditMode();
-  };
 
   const questionData = onlySelectedOption ? selectedQuestions : questionAPIData;
   if (!questionData) return;
@@ -83,7 +53,6 @@ const TabPanelItem: React.FC<TabPanelItemProps> = ({
       `}
     >
       <div
-        ref={tabContentRef}
         css={css`
           position: relative;
           display: flex;
@@ -104,52 +73,12 @@ const TabPanelItem: React.FC<TabPanelItemProps> = ({
         >
           <QuestionAddForm workbookId={workbook.workbookId} />
         </div>
-        <div
-          css={css`
-            display: flex;
-            flex-direction: column;
-            row-gap: 1.2rem;
-            height: 100%;
-            overflow-y: auto;
-            flex: 1;
-            padding: 1rem;
-          `}
-        >
-          {questionData.map((question) => (
-            <div
-              key={question.questionId}
-              onClick={() => handleQuestionChecked(question.questionId)}
-              css={css`
-                display: flex;
-                align-items: center;
-                column-gap: 0.5rem;
-              `}
-            >
-              {isEditMode && (
-                <CheckBox
-                  id={`question-${question.questionId}`}
-                  checked={isCheckedQuestion(question.questionId)}
-                  onInputChange={() =>
-                    handleQuestionChecked(question.questionId)
-                  }
-                />
-              )}
-              <QuestionSelectionBoxAccordion
-                key={question.questionId}
-                question={question}
-                workbookId={workbook.workbookId}
-                isSelectable={!isEditMode}
-              />
-            </div>
-          ))}
-        </div>
-        {isEditMode && (
-          <WorkbookEditModeDialog
-            count={checkQuestionCount()}
-            onCancelClick={handleCancelEditMode}
-            onDeleteClick={() => void handleDeleteQuestion()}
-          />
-        )}
+        <QuestionAccordionList
+          isEditMode={isEditMode}
+          cancelEditMode={() => setIsEditMode(false)}
+          questionData={questionData}
+          workbookId={workbook.workbookId}
+        />
         <div
           css={css`
             display: flex;
